+++ src/arenashooter/entities/Collider.java
@@ -1,16 +1,16 @@
 package arenashooter.entities;
 
-import arenashooter.engine.math.Vec2d;
+import arenashooter.engine.math.Vec2f;
 
 public class Collider extends Spatial {
-	Vec2d extent;
+	Vec2f extent;
 	
 	boolean testColl(Collider other) {
 		//TODO: faire, merci
 		return false;
 	}
 
-	public Collider(Vec2d extent) {
+	public Collider(Vec2f extent) {
 		super();
 		this.extent = extent;
 	}
