+++ src/arenashooter/entities/Plateform.java
@@ -1,12 +1,12 @@
 package arenashooter.entities;
 
-import arenashooter.engine.math.Vec2d;
+import arenashooter.engine.math.Vec2f;
 import arenashooter.engine.math.Vec3f;
 
 public class Plateform extends Spatial {
 	
 
-	public Plateform(Vec2d extent, Vec3f rgb) {
+	public Plateform(Vec2f extent, Vec3f rgb) {
 		Collider coll = new Collider(extent);
 		coll.extent = extent;
 		attachToParent(coll, "collider");
