+++ src/arenashooter/entities/Spatial.java
@@ -1,8 +1,8 @@
 package arenashooter.entities;
 
-import arenashooter.engine.math.Vec2d;
+import arenashooter.engine.math.Vec2f;
 
 public class Spatial extends Entity {
-	public Vec2d position = new Vec2d();
+	public Vec2f position = new Vec2f();
 	public float rotation = 0;
 }
